20211129 조환영
delivery 조회, seller 정보 수정

+++ src/main/java/com/team/controller/seller/SellerController.java
@@ -1,24 +1,20 @@
 package com.team.controller.seller;
 
-import java.net.URI;
+import java.sql.Timestamp;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
-import com.team.entity.Product;
 import com.team.entity.Seller;
 import com.team.jwt.JwtUtil;
 import com.team.service.SellerService;
+import com.team.vo.DeliveryVO;
 import com.team.vo.ProductVO;
 import com.team.vo.QnAVO;
 
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
-import org.springframework.http.ResponseEntity;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-import org.springframework.util.Assert;
-import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -38,39 +34,6 @@ public class SellerController {
     JwtUtil jwtUtil;
 
     int PAGECNT = 10;
-
-    @GetMapping(value = "/test")
-    public Map<String, Object> testttt(){
-
-        Map<String, Object> map = new HashMap<>();
-        // List<Seller> seller = sservice.selectSellerAll();
-        // Assert.notNull(seller1, "데이터를 담아서 보내주세요!");
-        Seller seller = new Seller();
-        System.out.println(seller.getRole().toString());
-
-        map.put("data", seller);
-        map.put("res", ResponseEntity.ok(seller));
-        // map.put("status", HttpStatus.OK);
-
-        return map;
-    }
-
-    // 판매자 계정 생성 페이지
-    @RequestMapping(value="/insert", method=RequestMethod.GET,
-    consumes = MediaType.ALL_VALUE, produces = MediaType.APPLICATION_JSON_VALUE)
-    public Map<String, Object> registSellerGet() {
-
-        Map<String, Object> map = new HashMap<>();
-        try{
-            map.put("200", "connect");
-        }
-        catch(Exception e){
-            e.printStackTrace();
-            map.put("400", e.hashCode());
-        }
-
-        return map;
-    }
     
     // 판매자 계정 생성
     // 127.0.0.1:8080/ROOT/seller/insert
@@ -82,7 +45,7 @@ public class SellerController {
             BCryptPasswordEncoder bcpe = new BCryptPasswordEncoder();
             seller.setSellerPw(bcpe.encode(seller.getSellerPw()));
             sservice.insertSeller(seller);
-            map.put("200", "connect");
+            map.put("status", 200);
         }
         catch(Exception e){
             e.printStackTrace();
@@ -91,24 +54,19 @@ public class SellerController {
         return map;
     }
 
-    // 판매자 정보 전달
+    // 판매자 정보 조회
+    // 127.0.0.1:8080/ROOT/seller/update
     @RequestMapping(value = "/update", method = RequestMethod.GET,
     consumes = MediaType.ALL_VALUE, produces = MediaType.APPLICATION_JSON_VALUE)
-    public Map<String, Object> getSellerInfo(){
+    public Map<String, Object> getSellerInfo(
+        @RequestHeader("token") String token){
         Map<String, Object> map = new HashMap<>();
         try{
-            // 토큰 생기면 수정할것 토큰에서 아이디 값 받아서 오는걸로 수정할것
-            Seller seller = new Seller();
-            String dbSellerPW = sservice.selectSellerOne(seller.getSellerId()).getSellerPw();
-            String inputSellerPW = seller.getSellerPw();
-            if(dbSellerPW == inputSellerPW){
-                Seller seller2 = sservice.selectSellerOne(seller.getSellerId());
-                map.put("seller", seller2);
-                map.put("200", "connect");
-            }
-            else{
-                map.put("warning", "wrong password");
-            }
+            String sellerid = jwtUtil.extractUsername(token); // token을 통해 판매자 정보 찾기
+            Seller oldSeller = sservice.selectSellerOne(sellerid);
+            map.put("seller", oldSeller);
+            map.put("status", 200);
+        
         }catch(Exception e){
             e.printStackTrace();
             map.put("400", e.hashCode());
@@ -117,6 +75,7 @@ public class SellerController {
     }
 
     // 판매자정보 수정
+    // 127.0.0.1:8080/ROOT/seller/update
     @RequestMapping(value = "/update", method = RequestMethod.PUT,
     consumes = MediaType.ALL_VALUE, produces = MediaType.APPLICATION_JSON_VALUE)
     public Map<String, Object> postSellerInfo(@RequestBody Seller seller,
@@ -125,69 +84,70 @@ public class SellerController {
         try{
             String sellerid = jwtUtil.extractUsername(token); // token을 통해 판매자 정보 찾기
             Seller oldSeller = sservice.selectSellerOne(sellerid);
-            // if(seller.getSellerEmail() != null){
-            //     oldSeller.setSellerEmail(seller.getSellerEmail());            
-            // }
             BCryptPasswordEncoder bcpe = new BCryptPasswordEncoder();
-            oldSeller.setSellerPw(bcpe.encode(seller.getSellerPw()));
-            oldSeller.setSellerEmail(seller.getSellerEmail());    
-            oldSeller.setSellerPhone(seller.getSellerPhone());
-            oldSeller.setStorePhone(seller.getStorePhone());
-            int updateCheck = sservice.updateSeller(oldSeller);
-            if(updateCheck == 1){
-                map.put("test", oldSeller.getSellerPw());
-                map.put("status", 200);
+            
+            System.out.println(sellerid.equals(oldSeller.getSellerId()));
+            System.out.println(bcpe.matches(seller.getSellerPw(), oldSeller.getSellerPw()));
+            System.out.println(seller.toString());
+            if(sellerid.equals(oldSeller.getSellerId()) && bcpe.matches(seller.getSellerPw(), oldSeller.getSellerPw())){
+                if(seller.getSellerNewPw() != null) {
+                    oldSeller.setSellerPw(bcpe.encode(seller.getSellerNewPw())); // 새로운 비밀번호
+                }
+                if(seller.getSellerEmail() !=  null){
+                    oldSeller.setSellerEmail(seller.getSellerEmail());  //메일
+                }
+                if(seller.getSellerPhone() != null){
+                    oldSeller.setSellerPhone(seller.getSellerPhone());  //연락처
+                }
+                if(seller.getStorePhone() != null){
+                    oldSeller.setStorePhone(seller.getStorePhone());    //매장 연락처
+                }
+                // System.out.println(oldSeller.toString());
+                Timestamp timestamp = new Timestamp(System.currentTimeMillis());
+                oldSeller.setSellerEditdate(timestamp); //수정 날짜
+                int updateCheck = sservice.updateSeller(oldSeller);
+                if(updateCheck == 1){
+                    map.put("status", 200);
+                }
             }
         }
         catch(Exception e){
             e.printStackTrace();
             map.put("400", e.hashCode());
         }
-
         return map;
     }
 
     // 판매자 계정 삭제
+    // 127.0.0.1:8080/ROOT/seller/delete
     @RequestMapping(value = "/delete", method = RequestMethod.DELETE,
     consumes = MediaType.ALL_VALUE, produces = MediaType.APPLICATION_JSON_VALUE)
-    public Map<String, Object> deleteSeller(Seller seller){
+    public Map<String, Object> deleteSeller(@RequestBody Seller seller,
+        @RequestHeader("token") String token){
         Map<String, Object> map = new HashMap<>();
-        // 토큰이랑 로그인 생기면 수정할것
         try{
-            sservice.deleteSeller(seller);
-            map.put("status", 200);
-
+            BCryptPasswordEncoder bcpe = new BCryptPasswordEncoder();
+            String sellerid = jwtUtil.extractUsername(token);
+            String sellerpw = sservice.selectSellerOne(sellerid).getSellerPw();
+            if(sellerid.equals(seller.getSellerId())){  //로그인 한 id와 입력한 id가 같은지 비교
+                if(bcpe.matches(seller.getSellerPw(), sellerpw)){
+                    sservice.deleteSeller(seller);
+                    map.put("status", 200);
+                }
+            }
+            else{
+                map.put("status", 300);
+            }
         }
         catch(Exception e){
             e.printStackTrace();
             map.put("400", e.hashCode());
         }
         return map;
-    }
-
-    // 판매자 계정 삭제 페이지
-    @RequestMapping(value = "/delete", method = RequestMethod.GET,
-    consumes = MediaType.ALL_VALUE, produces = MediaType.APPLICATION_JSON_VALUE)
-    public Map<String, Object> GetDeletePage(){
-        Map<String, Object> map = new HashMap<>();
-        try{
-            // 토큰이 있다면 진행하게 할것(if)
-            // 토큰이 있어서 진행이 됬다면 토큰에서 아이디값이 빼와서 화면에 보여줌
-            // 버튼 누르면 delete 수행되게 함
-            // DB Table에서 지워지면 자동으로 log table 에 저장되게함
-            map.put("status", 200);
-        }
-        catch(Exception e){
-            e.printStackTrace();
-            map.put("400", e.hashCode());
-        }
-
-        return map;
-
     }
 
     //판매자 별 물품 조회
-    // 127.0.0.1:8080/ROOT/seller/select_prolist?id= 판매자id&page=
+    // 127.0.0.1:8080/ROOT/seller/select_prolist?page=1
     @RequestMapping(value = "/select_prolist", method = {
         RequestMethod.GET }, consumes = MediaType.ALL_VALUE, produces = MediaType.APPLICATION_JSON_VALUE)
     public Map<String, Object> selectProductGET(
@@ -210,7 +170,7 @@ public class SellerController {
             }
             map.put("cnt", (cnt - 1) / 10 + 1); //페이지 수
             map.put("list", list);
-            map.put("result", 1);
+            map.put("result", 200);
         } catch (Exception e) {
             e.printStackTrace();
             map.put("result", e.hashCode());
@@ -219,11 +179,35 @@ public class SellerController {
     }
 
     //판매자 별 주문 조회
-
-    //판매자 별 배송 조회(송장번호 입력 용)
+    // 127.0.0.1:8080/ROOT/seller/select_dellist?page=1
+    @RequestMapping(value = "/select_dellist", method = {
+        RequestMethod.GET }, consumes = MediaType.ALL_VALUE, produces = MediaType.APPLICATION_JSON_VALUE)
+    public Map<String, Object> selectDelGET(
+    @RequestHeader("token") String token,
+    @RequestParam(value = "page", defaultValue = "1")int page) {
+        Map<String, Object> map = new HashMap<>();
+        try {
+            String sellerid = jwtUtil.extractUsername(token); // token을 통해 판매자 정보 찾기
+            long cnt = sservice.countBySelleridDelivery(sellerid); //sellerid 별 delivery 갯수
+            Map<String, Object> map1 = new HashMap<>();
+            int rpage1 = page * PAGECNT;
+            int rpage = rpage1 - (PAGECNT - 1);
+            map1.put("page", rpage);
+            map1.put("page1", rpage1);
+            map1.put("id", sellerid);
+            List<DeliveryVO> list = sservice.selectDelList(map1);
+            map.put("cnt", (cnt - 1) / 10 + 1); //페이지 수
+            map.put("list", list);
+            map.put("result", 200);
+        } catch (Exception e) {
+            e.printStackTrace();
+            map.put("result", e.hashCode());
+        }
+        return map;
+    }
 
     //판매자 별 qna 조회(qna 제품 별 조회)
-    // 127.0.0.1:8080/ROOT/seller/select_qnalist?page=
+    // 127.0.0.1:8080/ROOT/seller/select_qnalist?page=1
     @RequestMapping(value = "/select_qnalist", method = {
         RequestMethod.GET }, consumes = MediaType.ALL_VALUE, produces = MediaType.APPLICATION_JSON_VALUE)
     public Map<String, Object> selectQnaGET(
@@ -242,7 +226,7 @@ public class SellerController {
             List<QnAVO> list = sservice.selectQnaList(map1);
             map.put("cnt", (cnt - 1) / 10 + 1); //페이지 수
             map.put("list", list);
-            map.put("result", 1);
+            map.put("result", 200);
         } catch (Exception e) {
             e.printStackTrace();
             map.put("result", e.hashCode());
